Fix installation step

+++ setup.sh
@@ -5,7 +5,8 @@ umacro_version="v0.0.2"
 [[ -d /etc/umacro/ ]] || mkdir /etc/umacro
 [[ -f /etc/umacro/umacro_conf.yml ]] || touch /etc/umacro/umacro_conf.yml
 
-rm /usr/local/bin/umacro
+[[ -f /usr/local/bin/umacro ]] || rm /usr/local/bin/umacro
+
 wget -P /usr/local/bin https://github.com/dhsathiya/UMacro/releases/download/$umacro_version/umacro -q --show-progress
 
 chmod +x /usr/local/bin/umacro
